requirements.txt
# =============================================================================
# Talk2YourServer - Python Dependencies
# =============================================================================
# Install with: pip install -r requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# Core - Telegram Bot Framework
# -----------------------------------------------------------------------------
python-telegram-bot>=20.7

# -----------------------------------------------------------------------------
# HTTP Clients
# -----------------------------------------------------------------------------
httpx>=0.25.0
aiohttp>=3.9.0

# -----------------------------------------------------------------------------
# LLM Provider Clients
# -----------------------------------------------------------------------------
# Groq - Fast cloud inference
groq>=0.4.0

# OpenAI - GPT models
openai>=1.0.0

# Anthropic - Direct API access (optional, Claude Code uses CLI)
anthropic>=0.18.0

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
# Async PostgreSQL driver
asyncpg>=0.29.0

# -----------------------------------------------------------------------------
# Configuration
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0

# -----------------------------------------------------------------------------
# System Integration
# -----------------------------------------------------------------------------
# Systemd watchdog support (optional, for service mode)
sdnotify>=0.3.0

# Process management
psutil>=5.9.0

# -----------------------------------------------------------------------------
# Development (optional)
# -----------------------------------------------------------------------------
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# ruff>=0.1.0
